+++ libs/coin-modules/coin-aptos/src/network/client.ts
@@ -309,9 +309,12 @@ export class AptosAPI {
     });
 
     return Promise.all(
-      queryResponse.data.account_transactions.map(({ transaction_version }) => {
-        return this.richItemByVersion(transaction_version);
-      }),
+      queryResponse.data.account_transactions
+        .slice()
+        .sort((a, b) => b.transaction_version - a.transaction_version)
+        .map(({ transaction_version }) => {
+          return this.richItemByVersion(transaction_version);
+        }),
     );
   }