+++ libs/coin-modules/coin-stellar/src/api/index.ts
@@ -2,9 +2,13 @@ import {
   AlpacaApi,
   Block,
   BlockInfo,
+  Cursor,
+  Page,
   FeeEstimation,
   Operation,
   Pagination,
+  Stake,
+  Reward,
   TransactionIntent,
 } from "@ledgerhq/coin-framework/api/index";
 import coinConfig, { type StellarConfig } from "../config";
@@ -39,6 +43,12 @@ export function createApi(config: StellarConfig): AlpacaApi<StellarMemo> {
     },
     getBlockInfo(_height: number): Promise<BlockInfo> {
       throw new Error("getBlockInfo is not supported");
+    },
+    getStakes(_address: string, _cursor?: Cursor): Promise<Page<Stake>> {
+      throw new Error("getStakes is not supported");
+    },
+    getRewards(_address: string, _cursor?: Cursor): Promise<Page<Reward>> {
+      throw new Error("getRewards is not supported");
     },
   };
 }