+++ libs/ledger-live-common/src/families/stellar/ui.ts
@@ -1,15 +1,22 @@
-import { StellarOperationExtra } from "./types";
+import { StellarMemo, StellarOperationExtra } from "./types";
 
 /**
  * Format stellar memo value for display.
  */
 export const formatMemo = (extra: StellarOperationExtra): string | undefined => {
-  switch (extra?.memo?.type) {
+  const memo: StellarMemo | string | undefined = extra.memo;
+
+  // for backward compatibility for previous operation structure, as there is no
+  // data migration system in place
+  // noinspection SuspiciousTypeOfGuard
+  if (typeof memo == "string") return memo;
+
+  switch (memo?.type) {
     case "MEMO_ID":
     case "MEMO_TEXT":
     case "MEMO_HASH":
     case "MEMO_RETURN":
-      return extra?.memo.value;
+      return memo?.value;
   }
   return undefined;
 };