+++ libs/ledger-live-common/src/modularDrawer/utils/index.ts
@@ -1,6 +1,7 @@
 import { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
 import { CurrenciesByProviderId } from "../../deposit/type";
 import { haveOneCommonProvider } from "./haveOneCommonProvider";
+import { getBalanceAndFiatValue } from "./getBalanceAndFiatValue";
 
 function isCorrespondingCurrency(
   elem: CryptoOrTokenCurrency,
@@ -27,4 +28,9 @@ const getEffectiveCurrency = (
   return provider.currenciesByNetwork.find(elem => currencyIds.includes(elem.id)) ?? currency;
 };
 
-export { isCorrespondingCurrency, getEffectiveCurrency, haveOneCommonProvider };
+export {
+  isCorrespondingCurrency,
+  getEffectiveCurrency,
+  haveOneCommonProvider,
+  getBalanceAndFiatValue,
+};