+++ libs/ledger-services/trust/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ledgerhq/ledger-trust-service
 
+## 0.3.10
+
+### Patch Changes
+
+- Updated dependencies [[`75a08cc`](https://github.com/LedgerHQ/ledger-live/commit/75a08cc3061347bae98ddef7ac3cdcd6181ddab5)]:
+  - @ledgerhq/live-env@2.14.0
+  - @ledgerhq/live-network@2.0.15
+
+## 0.3.10-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`75a08cc`](https://github.com/LedgerHQ/ledger-live/commit/75a08cc3061347bae98ddef7ac3cdcd6181ddab5)]:
+  - @ledgerhq/live-env@2.14.0-next.0
+  - @ledgerhq/live-network@2.0.15-next.0
+
 ## 0.3.9
 
 ### Patch Changes