+++ libs/ledgerjs/packages/cryptoassets-evm-signatures/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/cryptoassets-evm-signatures
 
+## 13.5.12
+
+### Patch Changes
+
+- Updated dependencies [[`75a08cc`](https://github.com/LedgerHQ/ledger-live/commit/75a08cc3061347bae98ddef7ac3cdcd6181ddab5)]:
+  - @ledgerhq/live-env@2.14.0
+
+## 13.5.12-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`75a08cc`](https://github.com/LedgerHQ/ledger-live/commit/75a08cc3061347bae98ddef7ac3cdcd6181ddab5)]:
+  - @ledgerhq/live-env@2.14.0-next.0
+
 ## 13.5.11
 
 ### Patch Changes